[chore] #14 패키지 정리
