+++ app/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".LoginActivity"
+            android:name=".presentation.ui.login.LoginActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.NOWSOPTAndroid">
@@ -27,12 +27,12 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.ui.main.MainActivity"
             android:exported="false"
             android:label="@string/app_name"
             android:theme="@style/Theme.NOWSOPTAndroid"></activity>
         <activity
-            android:name=".SignupActivity"
+            android:name=".presentation.ui.signup.SignupActivity"
             android:exported="false"
             android:label="@string/app_name"
             android:theme="@style/Theme.NOWSOPTAndroid"></activity>
