+++ app/src/main/java/com/sopt/now/compose/presentation/model/User.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.presentation.model
 
 import androidx.annotation.DrawableRes
