+++ app/src/main/java/com/sopt/now/compose/presentation/ui/main/home/HomeScreen.kt
@@ -1,20 +1,12 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.presentation.ui.main.home
 
-import android.util.Log
 import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.sopt.now.compose.data.DTO.response.ResponseUserInfoDataDto
-import com.sopt.now.compose.data.DTO.response.ResponseUserInfoDto
-import com.sopt.now.compose.data.ServicePool
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+import com.sopt.now.compose.R
+import com.sopt.now.compose.presentation.component.UserItem
 
 @Composable
 fun HomeScreen(userId: String) {
