+++ app/src/main/java/com/sopt/now/compose/presentation/ui/main/mypage/MypageScreen.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.presentation.ui.main.mypage
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -21,6 +21,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.sopt.now.compose.R
 
 @Composable
 fun MypageScreen(id: String? = "", pw: String? = "", nickname: String? = "", mbti: String? = "") {
