+++ app/src/main/java/com/sopt/now/compose/presentation/ui/main/search/SearchScreen.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.presentation.ui.main.search
 
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
