+++ app/src/main/java/com/sopt/now/compose/presentation/ui/signup/SignupActivity.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.presentation.ui.signup
 
 import android.content.Intent
 import android.os.Bundle
@@ -23,34 +23,26 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight.Companion.Bold
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
 import androidx.compose.ui.text.input.PasswordVisualTransformation
 import androidx.compose.ui.text.input.VisualTransformation
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.lifecycle.MutableLiveData
-import com.sopt.now.compose.data.DTO.request.RequestLoginDto
-import com.sopt.now.compose.data.DTO.response.ResponseSignupDto
-import com.sopt.now.compose.data.Key.ID
-import com.sopt.now.compose.data.Key.MBTI
-import com.sopt.now.compose.data.Key.NICKNAME
-import com.sopt.now.compose.data.Key.PW
-import com.sopt.now.compose.data.ServicePool
-import com.sopt.now.compose.ui.theme.NOWSOPTAndroidTheme
-import retrofit2.Call
-import retrofit2.Callback
-import retrofit2.Response
+import com.sopt.now.compose.presentation.ui.login.LoginActivity
+import com.sopt.now.compose.R
+import com.sopt.now.compose.data.DTO.request.RequestSignupDto
+import com.sopt.now.compose.theme.NOWSOPTAndroidTheme
 
-class LoginActivity : ComponentActivity() {
+class SignupActivity : ComponentActivity() {
+    val signupViewModel = SignupViewModel()
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContent {
@@ -59,26 +51,30 @@ class LoginActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    Login(
-                        signupId = intent.getStringExtra(ID),
-                        signupPw = intent.getStringExtra(PW),
-                        nickname = intent.getStringExtra(NICKNAME),
-                        mbti = intent.getStringExtra(MBTI)
-                    )
+                    Signup(signupViewModel)
                 }
             }
+        }
+
+        initLiveData()
+    }
+
+    private fun initLiveData() {
+        signupViewModel.liveData.observe(this) {
+            Toast.makeText(this@SignupActivity, it.message, Toast.LENGTH_SHORT).show()
         }
     }
 }
 
+
+
 @Composable
-fun Login(signupId: String?, signupPw: String?, nickname: String?, mbti: String?) {
+fun Signup(signupViewModel: SignupViewModel) {
     val context = LocalContext.current
     var id by remember { mutableStateOf("") }
     var pw by remember { mutableStateOf("") }
-
-    val authService by lazy { ServicePool.authService }
-    val liveData = MutableLiveData<SignupState>()
+    var nickname by remember { mutableStateOf("") }
+    var phone by remember { mutableStateOf("") }
 
     var shouldShowPassword by remember {
         mutableStateOf(false)
@@ -97,9 +93,9 @@ fun Login(signupId: String?, signupPw: String?, nickname: String?, mbti: String?
             .fillMaxWidth()
     ) {
         Text(
-            text = stringResource(id = R.string.tv_login_title),
+            text = stringResource(id = R.string.tv_signup_title),
             fontSize = 30.sp,
-            fontWeight = Bold,
+            fontWeight = FontWeight.Bold,
             textAlign = TextAlign.Center,
             modifier = Modifier
                 .fillMaxWidth()
@@ -108,11 +104,11 @@ fun Login(signupId: String?, signupPw: String?, nickname: String?, mbti: String?
         Text(
             text = stringResource(id = R.string.tv_login_id),
             fontSize = 20.sp,
-            fontWeight = Bold,
+            fontWeight = FontWeight.Bold,
             textAlign = TextAlign.Start,
             modifier = Modifier
                 .fillMaxWidth()
-                .padding(top = 150.dp)
+                .padding(top = 50.dp)
         )
         TextField(
             value = id,
@@ -127,7 +123,7 @@ fun Login(signupId: String?, signupPw: String?, nickname: String?, mbti: String?
         Text(
             text = stringResource(id = R.string.tv_login_pw),
             fontSize = 20.sp,
-            fontWeight = Bold,
+            fontWeight = FontWeight.Bold,
             textAlign = TextAlign.Start,
             modifier = Modifier
                 .fillMaxWidth()
@@ -158,46 +154,53 @@ fun Login(signupId: String?, signupPw: String?, nickname: String?, mbti: String?
                 }
             }
         )
+        Text(
+            text = stringResource(id = R.string.tv_signup_nickname),
+            fontSize = 20.sp,
+            fontWeight = FontWeight.Bold,
+            textAlign = TextAlign.Start,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 20.dp)
+        )
+        TextField(
+            value = nickname,
+            onValueChange = {
+                nickname = it
+            },
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 10.dp),
+            label = { Text(stringResource(id = R.string.et_signup_nickname_hint)) }
+        )
+        Text(
+            text = stringResource(id = R.string.tv_signup_phone),
+            fontSize = 20.sp,
+            fontWeight = FontWeight.Bold,
+            textAlign = TextAlign.Start,
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 20.dp)
+        )
+        TextField(
+            value = phone,
+            onValueChange = {
+                phone = it
+            },
+            modifier = Modifier
+                .fillMaxWidth()
+                .padding(top = 10.dp),
+            label = { Text(stringResource(id = R.string.et_signup_phone_hint)) }
+        )
         Button(
             onClick = {
-                val requestLoginDto = RequestLoginDto(id, pw)
-                authService.login(requestLoginDto).enqueue(object : Callback<ResponseSignupDto> {
-                    override fun onResponse(
-                        call: Call<ResponseSignupDto>,
-                        response: Response<ResponseSignupDto>
-                    ) {
-                        if (response.isSuccessful) {
-                            val data: ResponseSignupDto? = response.body()
-                            val userId = response.headers()["location"]
-
-                            val intent = Intent(context, MainActivity::class.java)
-                            intent.putExtra("userId", userId)
-
-                            Toast.makeText(context, "$userId 님 로그인에 성공했습니다", Toast.LENGTH_SHORT).show()
-                            context.startActivity(intent)
-                        } else {
-                            liveData.value = SignupState(true, "아이디와 비밀번호가 일치하지 않습니다")
-                        }
-                    }
-
-                    override fun onFailure(call: Call<ResponseSignupDto>, t: Throwable) {
-                        liveData.value = SignupState(true, "서버 통신 에러")
-                    }
-                })
-
-
-                if (id == signupId && pw == signupPw) {
-                    Toast.makeText(context, R.string.login_success, Toast.LENGTH_SHORT).show()
-                    Intent(context, MainActivity::class.java).apply {
-                        putExtra(ID, id)
-                        putExtra(PW, pw)
-                        putExtra(NICKNAME, nickname)
-                        putExtra(MBTI, mbti)
-                        context.startActivity(this)
-                    }
-
+                val validationMsg = signupViewModel.checkSignupValidation(id, pw, nickname, phone)
+                if (validationMsg == R.string.signup_success) {
+                    val requestSignupDto = RequestSignupDto(id, pw, nickname, phone)
+                    signupViewModel.postSignup(requestSignupDto)
+                    context.startActivity(Intent(context, LoginActivity::class.java))
                 } else {
-                    Toast.makeText(context, R.string.login_fail, Toast.LENGTH_SHORT).show()
+                    Toast.makeText(context, validationMsg, Toast.LENGTH_SHORT).show()
                 }
             },
             colors = ButtonDefaults.buttonColors(
@@ -207,25 +210,7 @@ fun Login(signupId: String?, signupPw: String?, nickname: String?, mbti: String?
                 .fillMaxWidth()
                 .padding(top = 20.dp)
         ) {
-            Text(text = stringResource(id = R.string.btn_login))
-        }
-        Button(
-            onClick = {
-                val intent = Intent(context, SignupActivity::class.java)
-                context.startActivity(intent)
-            },
-            colors = ButtonDefaults.buttonColors(
-                containerColor = Color.Transparent
-            ),
-            modifier = Modifier.align(Alignment.End)
-        ) {
-            Text(text = stringResource(id = R.string.btn_signup), color = Color.Black)
+            Text(text = stringResource(id = R.string.btn_signup))
         }
     }
-}
-
-@Preview(showBackground = true)
-@Composable
-fun LoginPreview() {
-    Login(null, null, null, null)
 }
