+++ app/src/main/java/com/sopt/now/compose/presentation/ui/signup/SignupState.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose
+package com.sopt.now.compose.presentation.ui.signup
 
 data class SignupState(
     val state: Boolean,
