+++ app/src/main/java/com/sopt/now/compose/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.compose.ui.theme
+package com.sopt.now.compose.theme
 
 import androidx.compose.ui.graphics.Color
